feat/unify-folders-and-files-names (#972)
