+++ packages/i18n/lib/prepare-build.ts
@@ -1,4 +1,4 @@
-import setRelatedLocaleImports from './set_related_locale_import.js';
+import setRelatedLocaleImports from './set-related-locale-import.js';
 import { IS_DEV } from '@extension/env';
 import { cpSync, existsSync, mkdirSync } from 'node:fs';
 import { resolve } from 'node:path';
