+++ packages/module-manager/lib/zip-utils.ts
