+++ packages/storage/lib/base/index.ts
@@ -1,3 +1,3 @@
 export * from './base.js';
 export * from './enums.js';
-export type * from '../types.js';
+export type * from './types.js';
