+++ packages/storage/lib/index.ts
@@ -1,2 +1,2 @@
-export type { BaseStorage } from './base/index.js';
+export type * from './types.js';
 export * from './impl/index.js';
