+++ packages/ui/lib/components/error-display/ErrorDisplay.tsx
@@ -1,6 +1,6 @@
-import { ErrorHeader } from '@/lib/components/ErrorDisplay/ErrorHeader';
-import { ErrorResetButton } from '@/lib/components/ErrorDisplay/ErrorResetButton';
-import { ErrorStackTraceList } from '@/lib/components/ErrorDisplay/ErrorStackTraceList';
+import { ErrorHeader } from '@/lib/components/error-display/ErrorHeader';
+import { ErrorResetButton } from '@/lib/components/error-display/ErrorResetButton';
+import { ErrorStackTraceList } from '@/lib/components/error-display/ErrorStackTraceList';
 
 export const ErrorDisplay = ({ error, resetErrorBoundary }: { error?: Error; resetErrorBoundary?: () => void }) => (
   <div className="flex items-center justify-center bg-gray-50 py-6 px-4 sm:px-6 lg:px-8">
