+++ packages/ui/lib/components/index.ts
@@ -1,3 +1,3 @@
 export * from './ToggleButton';
 export * from './LoadingSpinner';
-export * from './ErrorDisplay/ErrorDisplay';
+export * from './error-display/ErrorDisplay';
