+++ packages/ui/lib/index.ts
@@ -1,3 +1,3 @@
 export * from './components/index';
 export * from './utils';
-export * from './withUI';
+export * from './with-ui';
