+++ pages/content-runtime/tsconfig.json
@@ -5,7 +5,8 @@
     "paths": {
       "@src/*": ["src/*"]
     },
-    "types": ["chrome", "node", "./global.d.ts"]
+    "types": ["chrome", "node", "./global.d.ts"],
+    "allowImportingTsExtensions": true
   },
   "include": ["src", "vite.config.mts"]
 }
