+++ pages/content/src/index.ts
@@ -1,4 +1,4 @@
-import { sampleFunction } from '@src/sampleFunction';
+import { sampleFunction } from '@src/sample-function';
 
 console.log('content script loaded');
